+++ lending/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v2.0.0] lending
 ### Releases
 - [PyPI v2.0.0] https://pypi.org/project/codat-lending/2.0.0 - lending
+
+## 2023-09-14 15:14:50
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.83.2 (2.109.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.1.0] lending
+### Releases
+- [PyPI v3.1.0] https://pypi.org/project/codat-lending/3.1.0 - lending